+++ packages/typegpu/src/shared/tseynit.ts
